apps/examples/fs_test: Unify examples for procfs test
Remove duplicated tests for procfs, proc_test folder and Update fs_test/procfs_test.c
