--- apps/examples/proc_test/Kconfig
@@ -1,15 +1,0 @@
-#
-# For a description of the syntax of this configuration file,
-# see kconfig-language at https://www.kernel.org/doc/Documentation/kbuild/kconfig-language.txt
-#
-
-config EXAMPLES_PROC_TEST
-	bool "\"Procfs\" example"
-	default n
-	depends on !DISABLE_MOUNTPOINT && FS_PROCFS
-	---help---
-		Enable the \"Procfs\" example
-
-config USER_ENTRYPOINT
-	string
-	default "proc_test_main" if ENTRY_PROC_TEST
