--- apps/examples/proc_test/Kconfig_ENTRY
@@ -1,3 +1,0 @@
-config ENTRY_PROC_TEST
-	bool "\"Procfs\" example"
-	depends on EXAMPLES_PROC_TEST
