--- apps/examples/proc_test/README.txt
@@ -1,13 +1,0 @@
-examples/proc_test
-^^^^^^^^^^^^^^^^^^
-
-  This is an sample example of proc usages. You can find informations about process, cpu usages, version
-  and sytem up time.
-
-  Configs (see the details on Kconfig):
-  * CONFIG_EXAMPLES_PROC_TEST
-
-  Depends on:
-  * CONFIG_DISABLE_MOUNTPOINT
-  * CONFIG_FS_PROCFS
-
